requirements.txt
# Vilabot.cat - Dependencies
# Python 3.11+ recommended

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Async HTTP client for scraping
httpx==0.28.1

# HTML parsing
beautifulsoup4==4.12.3
lxml==5.3.0

# OpenAI API
openai==1.58.1

# Environment management
python-dotenv==1.0.1

# Data validation (included with FastAPI but pinning version)
pydantic==2.10.4
